+++ docs/apidoc/examples/netem.py
@@ -30,6 +30,7 @@ tc = {
     "enable": True,
     "default_delay": "20ms",
     "default_rate": "1gbit",
+    "groups": ["control", "compute"]
 }
 
 netem = Netem(tc, roles=roles)