+++ enoslib/tests/unit/test_network_utils.py
@@ -1,4 +1,5 @@
 from enoslib.service.netem.netem import (
+    _build_commands,
     _build_ip_constraints,
     _expand_description,
     _generate_default_grp_constraints,
@@ -278,11 +279,19 @@ class TestMergeConstraints(EnosTest):
 
 
 class TestBuildIpConstraints(EnosTest):
-    def test_build_ip_constraints(self):
-        # role distribution
-        rsc = {"grp1": [Host("node1")], "grp2": [Host("node2")]}
-        # ips informations
-        ips = {
+    def setUp(self):
+        self.n1 = Host("node1")
+        self.n2 = Host("node2")
+        self.rsc = {"grp1": [self.n1], "grp2": [self.n2]}
+        constraint = {
+            "src": "grp1",
+            "dst": "grp2",
+            "rate": "10mbit",
+            "delay": "10ms",
+            "loss": "0.1%",
+        }
+        self.constraints = [constraint]
+        self.ips = {
             "node1": {
                 "all_ipv4_addresses": ["ip11", "ip12"],
                 "devices": [
@@ -300,28 +309,8 @@ class TestBuildIpConstraints(EnosTest):
                 "enos_devices": ["eth0", "eth1"],
             },
         }
-        # the constraints
-        constraint = {
-            "src": "grp1",
-            "dst": "grp2",
-            "rate": "10mbit",
-            "delay": "10ms",
-            "loss": "0.1%",
-        }
-        constraints = [constraint]
 
-        ips_with_tc = _build_ip_constraints(rsc, ips, constraints)
-        # tc rules are applied on the source only
-        self.assertTrue("tc" in ips_with_tc["node1"])
-        tcs = ips_with_tc["node1"]["tc"]
-        # one rule per dest ip and source device
-        self.assertEqual(2 * 2, len(tcs))
-
-    def test_build_ip_constraints_bridge(self):
-        # role distribution
-        rsc = {"grp1": [Host("node1")], "grp2": [Host("node2")]}
-        # ips informations
-        ips = {
+        self.ips_with_bridge = {
             "node1": {
                 "all_ipv4_addresses": ["ip11", "ip12"],
                 "devices": [
@@ -349,17 +338,22 @@ class TestBuildIpConstraints(EnosTest):
                 "enos_devices": ["br0"],
             },
         }
-        # the constraints
-        constraint = {
-            "src": "grp1",
-            "dst": "grp2",
-            "rate": "10mbit",
-            "delay": "10ms",
-            "loss": "0.1%",
-        }
-        constraints = [constraint]
 
-        ips_with_tc = _build_ip_constraints(rsc, ips, constraints)
+    def test_build_ip_constraints(self):
+        ips_with_tc = _build_ip_constraints(self.rsc, self.ips, self.constraints)
+        # tc rules are applied on the source only
+        self.assertFalse("node2" in ips_with_tc)
+        # devices
+        self.assertTrue("devices" in ips_with_tc["node1"])
+        self.assertEqual(2, len(ips_with_tc["node1"]["devices"]))
+        self.assertTrue("tc" in ips_with_tc["node1"])
+        tcs = ips_with_tc["node1"]["tc"]
+        # one rule per dest ip and source device
+        self.assertEqual(2 * 2, len(tcs))
+
+    def test_build_ip_constraints_bridge(self):
+        ips_with_tc = _build_ip_constraints(
+            self.rsc, self.ips_with_bridge, self.constraints)
         # tc rules are applied on the source only
         self.assertTrue("tc" in ips_with_tc["node1"])
         tcs = ips_with_tc["node1"]["tc"]
@@ -370,3 +364,21 @@ class TestBuildIpConstraints(EnosTest):
         for tc in tcs:
             devices.add(tc["device"])
         self.assertCountEqual(["eth0"], list(devices))
+
+    def test_build_commands(self):
+        self.constraints[0]["symetric"] = True
+        ips_with_tc = _build_ip_constraints(
+            self.rsc, self.ips, self.constraints)
+        remove, add, rate, delay, filtr = _build_commands(ips_with_tc)
+        self.assertEqual(2, len(remove[self.n1]))
+        # symetric cases are handled prior to _build_ip_constraints
+        # so are we aren't symetric unless we add a symetric constraint
+        self.assertEqual(0, len(remove[self.n2]))
+        self.assertEqual(2, len(add[self.n1]))
+        self.assertEqual(0, len(add[self.n2]))
+        # we have as many rate class as possible (device, dest)
+        self.assertEqual(2 * 2, len(rate[self.n1]))
+        # delay
+        self.assertEqual(2 * 2, len(delay[self.n1]))
+        # we put filter on every (device, dest) possible
+        self.assertEqual(2 * 2, len(filtr[self.n1]))